requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
pydantic==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
PyJWT[crypto]==2.9.0
httpx==0.27.2
redis==5.1.1
alembic==1.13.3
bcrypt==4.2.1
python-multipart==0.0.12
